Move validation to output, add open-sdg-specific checks
