+++ sdg/outputs/OutputBase.py
@@ -1,14 +1,18 @@
 class OutputBase:
     """Base class for destinations of SDG data/metadata."""
 
-    def __init__(self, inputs, output_folder=''):
+
+    def __init__(self, inputs, schema, output_folder=''):
         """Constructor for OutputBase."""
         self.indicators = self.merge_inputs(inputs)
+        self.schema = schema
         self.output_folder = output_folder
+
 
     def execute():
         """Write the SDG output to disk."""
         raise NotImplementedError
+
 
     def merge_inputs(self, inputs):
         """Take the results of many inputs and merge into a single dict of indicators."""
@@ -25,3 +29,13 @@ class OutputBase:
                     merged_indicators[inid].set_meta(input.indicators[inid].meta)
 
         return merged_indicators
+
+
+    def validate(self):
+        """Validate the data and metadata for the indicators."""
+
+        status = True
+        for inid in self.indicators:
+            status = status & self.schema.validate(self.indicators[inid])
+
+        return status
