+++ sdg/schemas/SchemaInputBase.py
@@ -44,7 +44,6 @@ class SchemaInputBase:
                 print('Validation errors for indicator ' + indicator.inid)
                 for error in self.validator.iter_errors(indicator.meta):
                     print(error.message)
-                    print(error.schema)
 
         return status
